+++ src/scandeval/dataset_configs.py
@@ -182,7 +182,7 @@ SST5_CONFIG = DatasetConfig(
     huggingface_id="ScandEval/sst5-mini",
     task=SENT,
     languages=[EN],
-    prompt_prefix="The following are texts are their sentiment, which can be "
+    prompt_prefix="The following are texts and their sentiment, which can be "
     "'positive', 'neutral' or 'negative'.",
     prompt_template="Text: {text}\nSentiment: {label}",
     prompt_label_mapping=dict(
@@ -940,6 +940,22 @@ NO_SAMMENDRAG_CONFIG = DatasetConfig(
     "ovennevnte artikkelen.",
     num_few_shot_examples=1,
     max_generated_tokens=256,
+)
+
+NORGLM_MULTI_SUM = DatasetConfig(
+    name="norglm-multi-sum",
+    pretty_name="the summarisation part of the Norwegian NorGLM multi-task human annotated dataset "
+    "NO-Multi-QA-Sum",
+    huggingface_id="ScandEval/norglm-multi-sum",
+    task=SUMM,
+    languages=[NB, NN, NO],
+    prompt_prefix="Her følger nyhetsartikler med tilhørende sammendrag.",
+    prompt_template="Nyhetsartikkel: {text}\nSammendrag: {target_text}",
+    instruction_prompt="Nyhetsartikkel: {text}\n\nSkriv et sammendrag av den "
+    "ovennevnte artikkelen.",
+    num_few_shot_examples=1,
+    max_generated_tokens=256,
+    unofficial=True,
 )
 
 WIKI_LINGUA_NL_CONFIG = DatasetConfig(