+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Added the summarisation part of the Norwegian NorGLM multi-task human annotated
+  dataset NO-Multi-QA-Sum (`norglm-multi-sum`). This dataset is part of the NLEBench
+  Norwegian benchmarks. It has been marked as `unofficial` for now.
+
+## [Changelog]
+### Fixed
+- Small typo in prefix prompt used for few-shot evaluation of the English sentiment
+  classification dataset SST5.
+
 
 ## [Unreleased]
 ### Added